Create errors folders for custom error.
