+++ src/errors/400.js
@@ -1,0 +1,2 @@
+module.exports = (err, req, res, next) =>
+  res.status(400).send({ error: err.message });
